Move bracketed paste functionality to Graphics.Vty.Input.Paste
