+++ src/Graphics/Vty/Input/Classify.hs
@@ -13,11 +13,12 @@ module Graphics.Vty.Input.Classify
 
 import Graphics.Vty.Input.Events
 import Graphics.Vty.Input.Mouse
+import Graphics.Vty.Input.Paste
 import Graphics.Vty.Input.Classify.Types
 
 import Codec.Binary.UTF8.Generic (decode)
 
-import Data.List (inits, isPrefixOf, isInfixOf)
+import Data.List (inits)
 import qualified Data.Map as M( fromList, lookup )
 import Data.Maybe ( mapMaybe )
 import qualified Data.Set as S( fromList, member )
@@ -83,31 +84,3 @@ utf8Length c
     | c < 0xE0 = 2
     | c < 0xF0 = 3
     | otherwise = 4
-
-bracketedPasteStart :: String
-bracketedPasteStart = "\ESC[200~"
-
-bracketedPasteEnd :: String
-bracketedPasteEnd = "\ESC[201~"
-
-bracketedPasteStarted :: String -> Bool
-bracketedPasteStarted = isPrefixOf bracketedPasteStart
-
-bracketedPasteFinished :: String -> Bool
-bracketedPasteFinished = isInfixOf bracketedPasteEnd
-
-takeUntil :: (Eq a) => [a] -> [a] -> ([a],[a])
-takeUntil [] _ = ([], [])
-takeUntil cs sub
-  | length cs < length sub      = (cs, [])
-  | take (length sub) cs == sub = ([], drop (length sub) cs)
-  | otherwise                   = let (pre, suf) = takeUntil (tail cs) sub
-                                  in (head cs:pre, suf)
-
-parseBracketedPaste :: String -> KClass
-parseBracketedPaste s =
-    let (p, rest) = takeUntil (drop (length bracketedPasteStart) s) bracketedPasteEnd
-        rest' = if bracketedPasteEnd `isPrefixOf` rest
-                then drop (length bracketedPasteEnd) rest
-                else rest
-    in Valid (EvPaste p) rest'
