+++ vty.cabal
@@ -98,6 +98,7 @@ library
                        Graphics.Vty.Input.Classify.Parse
                        Graphics.Vty.Input.Loop
                        Graphics.Vty.Input.Mouse
+                       Graphics.Vty.Input.Paste
                        Graphics.Vty.Input.Terminfo
                        Graphics.Vty.PictureToSpans
                        Graphics.Vty.Span
